requirements.txt
streamlit
numpy 
pandas
plotly
tabula-py